requirements.txt
astroid==2.2.5
colorama==0.4.1
comtypes==1.1.7
decorator==4.4.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
prompt-toolkit==1.0.16
py==1.8.0
Pygments==2.4.2
pylint==2.3.1
pywebio==1.2.3
pywin32==301
pywinauto==0.6.6
retry==0.9.2
retrying==1.3.3
robotframework==4.1
robotframework-debuglibrary==1.1.4
robotframework-pabot==0.85
robotremoteserver==1.1
selenium==3.141.0
six==1.12.0
tornado==6.1
typed-ast==1.4.3
ua-parser==0.10.0
urllib3==1.25.3
user-agents==2.2.0
wcwidth==0.1.7
wrapt==1.11.2
